[SML] Write JSON type attribute for Reference links
When parsing the "typeOf" and "attachedTo" properties (which result in
a Reference object), readLinkProperties() puts the "type" attribute in
"RemoteSchema". However, when serializing these again, these would
always use a fixed "application/sml+json" type (as passed in the
writeLink mediaType parameter).

Now, the RemoteSchema is used in serialization if set, otherwise falling
back to the mediaType passed in.